+++ CHANGELOG.md
@@ -9,6 +9,8 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+## [v1.18.0](https://github.com/allenai/beaker-py/releases/tag/v1.18.0) - 2023-03-07
+
 ### Added
 
 - Added missing `host_networking`, `replicas`, and `leader_selection` fields to `TaskSpec`.